Updated SQLiteObject
Added DB handle to be internal, mask it out of props that are saved to
the database.
Working on Foreign Key flag, may need to be internal to Godot, instead
of relying on SQLite database.
Changed signature of _setup() as klass is not needed.
Added _populate_object, to translate SQLite data to Godot Object.
Implemented _find_one(), _find_many() and _all() functions.
Implemented exists() for the object.
Implemented save_database()
Implemented delete()
Added _to_string() override to print out the Class Name, Table Name, and
Primary key, if one exists, otherwise we use the instance id.