Updated SQLiteObject
- Renamed GodotTypes to BaseTypes
- Added _registry to keep track of scripts that implement SQLiteObject,
with String Keys for the global names (As used in Property Info
Dictionary)
- Added diagnostic print_registered_classes() function.
- Added diagnostic print_data_structure() function.
- At setup(), we no longer create a new object, instead expect the Class
we are setting up.  Introspection is handled through the GDScript class,
instead of the instance class.  Also added registering the class with
the registry.
- Changed calling of _setup() virtual from internal call (Always on
SQLiteObject, instead of script) to klass._setup()
- Changed how Flags.AUTO_INCREMENT determination is handled, to ensure
the field we are working on, is of type int or real.
- Updated _populate_object() to do lazy loading of Object types, will
look to see what class is being used, and if it is a part of our
registry, if so, load it up by fetching the data, and storing it in the
variable, otherwise if it isn't registered, using bytes_to_var() to
restore the object.
- Added check to functions to see if _db has been setup yet or not.
- Added check to save() and delete() to see if the instance we are
working with, has been recording in the database or not.  If not we
fail, otherwise we handle saving/deleting to/from the database.